[Feature #2] Refactor room class file
-Remove unneccesary files and place related class files into one file
